* commit '4c4b428628cace7cb3743944a13e4b0db969f225':
  some changes
  some changes
  add test telesalesResult
  rename ScoringOrderRequest
  refactor field and class names
  refactor build gradle
  refactor bff fields

+++ src/main/java/az/kapitalbank/bffumico/client/ordermarketplace/model/OrderProductItem.java
@@ -1,5 +1,7 @@
 package az.kapitalbank.bffumico.client.ordermarketplace.model;
 
+import java.math.BigDecimal;
+
 import lombok.AccessLevel;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
@@ -7,7 +9,7 @@ import lombok.experimental.FieldDefaults;
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class OrderProductItem {
-    Integer productAmount;
+    BigDecimal productAmount;
     String productId;
     String productName;
     String orderNo;
